backend/requirements.txt
PyJWT
boto3
ping3
Flask
pytest
twilio
aiohttp
asyncio
psycopg2
gunicorn
sqlalchemy
flask_cors
pytest-flask
cryptography
flask_migrate
python-dotenv
email-validator
Werkzeug==2.2.2
Flask-JWT-Extended
